requirements.txt
# Django Core
Django
djangorestframework

# PostgreSQL Driver
psycopg2-binary

# OCR and Image Processing
pytesseract
Pillow
djangorestframework-api-key
